+++ src/javascript/binary/websocket_pages/cashier/cashier.js
@@ -42,40 +42,10 @@ var Cashier = (function() {
     };
 
     return {
-        lock_cashier: lock_cashier,
         check_locked: check_locked,
         check_virtual_top_up: check_virtual_top_up
     };
 }());
-
-pjax_config_page("/cashier", function(){
-    return {
-        onLoad: function() {
-          if (!/\/cashier\.html/.test(window.location.pathname) || !page.client.is_logged_in) {
-              return;
-          } else {
-              Cashier.check_locked();
-              Cashier.check_virtual_top_up();
-              page.contents.topbar_message_visibility(TUser.get().landing_company);
-          }
-        }
-    };
-});
-
-pjax_config_page("/cashier/payment_methods", function(){
-    return {
-        onLoad: function() {
-            if (japanese_client()) {
-                window.location.href = page.url.url_for('/');
-            }
-            if (!page.client.is_logged_in || page.client.is_virtual()) {
-                return;
-            } else {
-                Cashier.check_locked();
-            }
-        }
-    };
-});
 
 module.exports = {
     Cashier: Cashier,
