+++ src/javascript/binary/websocket_pages/cashier/payment_agent_listws.js
@@ -146,31 +146,6 @@ var PaymentAgentListWS = (function() {
     };
 }());
 
-
-
-pjax_config_page("payment_agent_listws", function() {
-    return {
-        onLoad: function() {
-            BinarySocket.init({
-                onmessage: function(msg) {
-                    var response = JSON.parse(msg.data);
-                    if (response) {
-                        if (response.msg_type === "paymentagent_list") {
-                            PaymentAgentListWS.responseHandler(response);
-                        }
-                    }
-                    else {
-                        console.log('some error occured');
-                    }
-                }
-            });
-
-            Content.populate();
-            PaymentAgentListWS.init();
-        }
-    };
-});
-
 module.exports = {
     PaymentAgentListWS: PaymentAgentListWS,
 };
